Extract method decl parser
