+++ crates/mockiato-codegen/src/parse.rs
@@ -2,6 +2,7 @@ use crate::diagnostic::DiagnosticBuilder;
 use crate::result::Result;
 use proc_macro2::Span;
 pub(crate) mod method_decl;
+pub(crate) mod method_decl_parser;
 pub(crate) mod method_inputs;
 pub(crate) mod method_inputs_parser;
 pub(crate) mod mockable_attr_parser;
